+++ editors/helix/tree-sitter-mcc/queries/highlights.scm
@@ -67,6 +67,7 @@
   "else"
   "case"
   "when"
+  "with"
 ] @keyword.control.conditional
 
 [