+++ editors/neovim/queries/mcc/highlights.scm
@@ -78,6 +78,7 @@
   "else"
   "case"
   "when"
+  "with"
 ] @keyword.conditional
 
 [